# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Description of ARIS backend API",
        "title": "ARIS backend API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/public/popular-users": {
            "get": {
                "description": "Get public suggested users",
                "tags": [
                    "feed"
                ],
                "summary": "Get public suggested users",
                "operationId": "get-pub-pop-users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.suggestedUsersResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/latest-events": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Get latest events",
                "tags": [
                    "feed"
                ],
                "summary": "Get latest events",
                "operationId": "get-latest-events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.suggestedUsersResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/suggested": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Get suggested users",
                "tags": [
                    "feed"
                ],
                "summary": "Get suggested users",
                "operationId": "get-sug-users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.suggestedUsersResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommonErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CommonErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "user.suggestedUserDTO": {
            "type": "object",
            "properties": {
                "avatarLink": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.suggestedUsersResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.suggestedUserDTO"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}